move Promise into its own module
