+++ dnp3/src/master/extract.rs
@@ -91,7 +91,7 @@ mod test {
     use crate::app::control::CommandStatus;
     use crate::app::parse::parser::HeaderCollection;
     use crate::app::*;
-    use crate::master::handler::{HeaderInfo, ReadHandler};
+    use crate::master::{HeaderInfo, ReadHandler};
 
     use super::*;
