+++ dnp3/src/master/messages.rs
@@ -3,8 +3,8 @@ use crate::decode::DecodeLevel;
 use crate::link::EndpointAddress;
 use crate::master::error::PollError;
 use crate::master::error::{AssociationError, TaskError};
-use crate::master::handler::Promise;
 use crate::master::poll::PollMsg;
+use crate::master::promise::Promise;
 use crate::master::tasks::Task;
 use crate::master::{AssociationConfig, AssociationHandler, AssociationInformation, ReadHandler};
 use crate::transport::FragmentAddr;
