+++ dnp3/src/master/mod.rs
@@ -17,6 +17,7 @@ pub(crate) mod convert;
 pub(crate) mod extract;
 pub(crate) mod messages;
 pub(crate) mod poll;
+pub(crate) mod promise;
 pub(crate) mod task;
 pub(crate) mod tasks;
