+++ dnp3/src/master/tasks/file/close.rs
@@ -3,9 +3,10 @@ use crate::app::format::WriteError;
 use crate::app::parse::free_format::FreeFormatVariation;
 use crate::app::parse::parser::{HeaderDetails, Response};
 use crate::app::{FileStatus, FunctionCode, Group70Var4};
+use crate::master::promise::Promise;
 use crate::master::tasks::file::REQUEST_ID;
 use crate::master::tasks::{AppTask, NonReadTask, Task};
-use crate::master::{FileError, FileHandle, Promise, TaskError};
+use crate::master::{FileError, FileHandle, TaskError};
 
 pub(crate) struct CloseFileTask {
     pub(crate) handle: FileHandle,
