+++ dnp3/src/master/tasks/mod.rs
@@ -5,8 +5,8 @@ use crate::app::ResponseHeader;
 use crate::master::association::Association;
 use crate::master::error::TaskError;
 use crate::master::extract::extract_measurements;
-use crate::master::handler::Promise;
 use crate::master::poll::Poll;
+use crate::master::promise::Promise;
 use crate::master::request::{Classes, EventClasses};
 use crate::master::tasks::auto::AutoTask;
 use crate::master::tasks::command::CommandTask;
