+++ dnp3/src/master/tasks/read.rs
@@ -1,6 +1,6 @@
 use crate::app::format::write::HeaderWriter;
 use crate::master::error::TaskError;
-use crate::master::handler::Promise;
+use crate::master::promise::Promise;
 use crate::master::request::ReadRequest;
 use crate::master::tasks::{AppTask, ReadTask, Task};
 use crate::master::ReadHandler;
